fix: rollback file rename
