+++ aidial_adapter_openai/exceptions/openai.py
